+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Use nanoseconds since epoch in timestamp
 - Adds `get.from_directory` and `get.by_id` to query
+- Introduces stable IDs based on the commit-context
 
 # v4.8.4